Keep scoped path helpers separate from file system types
