+++ front/lib/api/sandbox_functions/slug.ts
@@ -1,7 +1,9 @@
-import { resolveCanonicalScopedPath } from "@app/lib/api/files/mount_path";
+import {
+  parseCanonicalScopedPath,
+  resolveCanonicalScopedPath,
+} from "@app/lib/api/files/mount_path";
 import { normalizeAppPrefix } from "@app/types/api/pod_function_reference";
 import { SANDBOX_FUNCTION_SLUG_SEGMENT_REGEX } from "@app/types/api/sandbox_functions";
-import { parseCanonicalScopedPath } from "@app/types/file_system";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
