+++ front/lib/resources/file_resource.ts
@@ -15,6 +15,7 @@ import {
   getConversationFilesBasePath,
   getPodFilesBasePath,
   getPodSandboxFunctionsBasePath,
+  isLegacyScopedPath,
   makeProcessedMountFileName,
   resolveCanonicalScopedPath,
 } from "@app/lib/api/files/mount_path";
@@ -64,7 +65,6 @@ import { renderLightWorkspaceType } from "@app/lib/workspace";
 import logger from "@app/logger/logger";
 import tracer from "@app/logger/tracer";
 import { CoreAPI } from "@app/types/core/core_api";
-import { isLegacyScopedPath } from "@app/types/file_system";
 import type {
   AuthorizedFileAccessAllowlist,
   AuthorizedFileRef,
